+++ packages/ui/Slottable/tests/useSlots-specs.js
@@ -1,4 +1,3 @@
-/* globals console */
 /* eslint no-console: ["error", { allow: ["warn", "error"] }] */
 
 import React from 'react';