+++ packages/ui/Touchable/Flick.js
@@ -23,7 +23,16 @@ class Flick {
 		this.onFlick = onFlick;
 
 		this.move(coords);
-	}
+	};
+
+	// This method will get the `onFlick` props in the existing `flickConfig` values.
+	updateProps = ({onFlick}) => {
+		// Check `tracking` gesture is not in progress. Check if gesture exists before updating the references to the `flickConfig`
+		if (!this.tracking) return;
+
+		// This will update the `flickConfig` with the new value
+		this.flickConfig.onFlick = onFlick;
+	};
 
 	move = ({x, y}) => {
 		if (!this.tracking) return;
@@ -38,15 +47,15 @@ class Flick {
 		if (this.moves.length > this.maxMoves) {
 			this.moves.shift();
 		}
-	}
+	};
 
 	blur = () => {
 		this.end();
-	}
+	};
 
 	cancel = () => {
 		this.tracking = false;
-	}
+	};
 
 	cancelJob = new Job(this.cancel);
 
@@ -94,7 +103,7 @@ class Flick {
 		}
 
 		this.tracking = false;
-	}
+	};
 }
 
 const defaultFlickConfig = {