+++ packages/ui/ViewManager/tests/ViewManager-specs.js
@@ -434,4 +434,38 @@ describe('ViewManager', () => {
 
 		expect(spy).toHaveBeenLastCalledWith({index: 0, previousIndex: 1});
 	});
+
+	test('should pass `rtl` prop to arranger when `true`', () => {
+		const spy = jest.spyOn(MockArranger, 'stay');
+		mount(
+			<ViewManager arranger={MockArranger} index={0} rtl>
+				<div key="view1">View 1</div>
+				<div key="view2">View 2</div>
+			</ViewManager>
+		);
+
+		const expected = {rtl: true};
+		const actual = spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+
+		spy.mockRestore();
+	});
+
+	test('should pass `rtl` prop to arranger when unset', () => {
+		const spy = jest.spyOn(MockArranger, 'stay');
+		mount(
+			<ViewManager arranger={MockArranger} index={0}>
+				<div key="view1">View 1</div>
+				<div key="view2">View 2</div>
+			</ViewManager>
+		);
+
+		const expected = {rtl: false};
+		const actual = spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+
+		spy.mockRestore();
+	});
 });